requirements.txt
pygame
-e git+https://gitlab.com/lama-corp/graphalama.git@dev#egg=graphalama
pyconfiglib
